[Fix #283] 이미지 초기화 및 버튼 비활성화
